first stab at fixing the arrow key problem in the reader view (must disable keyboard event listener when side menus are open, or dialog (still need some work to get this last one working))
